upd Diary/component
