+++ src/components/Diary/StatsBlock/StatisticsBlock.styled.js
@@ -1,0 +1,161 @@
+import styled from 'styled-components';
+
+export const ListStats = styled.ul`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    height: 340px;
+    gap: 20px 13px;
+    align-content: flex-start;
+  }
+
+  @media screen and (min-width: 768px) {
+    flex-direction: column;
+    height: 248px;
+    gap: 16px;
+  }
+
+  @media screen and (min-width: 1440px) {
+    flex-direction: row;
+    height: 380px;
+  }
+`;
+export const ItemStats = styled.li`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    width: 157px;
+    height: 96px;
+    padding: 12px;
+
+    border-radius: 12px;
+    border: 1px solid rgba(239, 237, 232, 0.2);
+    background-color: rgba(239, 237, 232, 0.05);
+
+    &:nth-of-type(even) {
+      width: 165px;
+    }
+
+    &:nth-child(5) {
+      height: 108px;
+    }
+
+    &:nth-child(6) {
+      height: 108px;
+    }
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 187px;
+    height: 116px;
+    padding: 18px;
+
+    &:nth-of-type(even) {
+      width: 187px;
+    }
+
+    &:nth-child(5) {
+      height: 116px;
+    }
+
+    &:nth-child(6) {
+      height: 116px;
+    }
+  }
+`;
+export const ItemContent = styled.div`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 28px;
+
+    &.last-item {
+      gap: 40px;
+    }
+  }
+
+  @media screen and (min-width: 768px) {
+    &.last-item {
+      gap: 28px;
+    }
+  }
+`;
+export const TextContent = styled.p`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    align-items: center;
+
+    font-feature-settings:
+      'clig' off,
+      'liga' off;
+    font-family: Roboto;
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 133.333%;
+
+    color: rgba(239, 237, 232, 0.4);
+
+    &.last-text {
+      line-height: 150%;
+    }
+  }
+
+  @media screen and (min-width: 768px) {
+    line-height: 150%;
+  }
+`;
+export const TotalContent = styled.p`
+  @media screen and (min-width: 375px) {
+    font-feature-settings:
+      'clig' off,
+      'liga' off;
+    font-family: Roboto;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 111.111%;
+
+    color: var(--primary-text-color);
+  }
+
+  @media screen and (min-width: 768px) {
+    font-size: 24px;
+    line-height: 133.333%;
+  }
+`;
+export const BlockWarning = styled.div`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    margin-top: 20px;
+  }
+
+  @media screen and (min-width: 768px) {
+    margin-top: 32px;
+  }
+`;
+export const TextWarning = styled.p`
+  @media screen and (min-width: 375px) {
+    width: 303px;
+
+    font-family: Roboto;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 128.571%;
+
+    color: var(--secondary-text-color);
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 561px;
+
+    font-size: 16px;
+    line-height: 150%;
+  }
+
+  @media screen and (min-width: 1440px) {
+    width: 358px;
+  }
+`;
