+++ src/components/Diary/TitleCallendar/TitleCall.styled.js
@@ -1,0 +1,136 @@
+import styled from 'styled-components';
+import DatePicker from 'react-datepicker';
+import './react-datepicker.css';
+
+export const BlockTitleCall = styled.div`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    align-items: center;
+    flex-direction: row-reverse;
+    justify-content: space-between;
+
+    margin-bottom: 40px;
+  }
+
+  @media screen and (min-width: 768px) {
+    flex-direction: column;
+    align-items: stretch;
+
+    margin-bottom: 32px;
+  }
+`;
+export const BlockCalendar = styled.div`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    gap: 8px 20px;
+  }
+`;
+export const CustomDatePicker = styled(DatePicker)`
+  @media screen and (min-width: 375px) {
+    width: 100px;
+
+    font-family: Roboto;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 111.111%;
+
+    background-color: transparent;
+    border: none;
+    color: var(--primary-text-color);
+
+    &:focus {
+      outline: none;
+    }
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 131px;
+
+    font-size: 24px;
+    line-height: 133.333%;
+  }
+`;
+export const BlockBtnDate = styled.div`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+`;
+export const BtnLeft = styled.button`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: transparent;
+    border: none;
+    padding: 0;
+  }
+
+  @media screen and (min-width: 768px) {
+    margin-right: 6px;
+  }
+`;
+export const BtnRight = styled.button`
+  @media screen and (min-width: 375px) {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: transparent;
+    border: none;
+    padding: 0;
+  }
+`;
+export const Title = styled.h2`
+  @media screen and (min-width: 375px) {
+    font-family: Roboto;
+    font-size: 24px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 116.667%;
+
+    color: var(--primary-text-color);
+  }
+
+  @media screen and (min-width: 768px) {
+    font-size: 32px;
+    line-height: 137.5%;
+  }
+`;
+export const SvgIconCalendar = styled.svg`
+  @media screen and (min-width: 375px) {
+    width: 20px;
+    height: 20px;
+    stroke: #ef8964;
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 24px;
+    height: 24px;
+  }
+`;
+export const SvgIconLeft = styled.svg`
+  @media screen and (min-width: 375px) {
+    width: 16px;
+    height: 16px;
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 20px;
+    height: 20px;
+  }
+`;
+export const SvgIconRight = styled.svg`
+  @media screen and (min-width: 375px) {
+    width: 16px;
+    height: 16px;
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 20px;
+    height: 20px;
+  }
+`;
